Clippy & Fmt

+++ tests/unavailable_prover.rs
@@ -9,7 +9,7 @@ async fn unavailable_prover() -> anyhow::Result<()> {
     init_tracing_subscriber();
     info!("Starting unavailable prover test");
 
-    let mut ref_tree = PoseidonTree::new(DEFAULT_TREE_DEPTH as usize + 1, ruint::Uint::ZERO);
+    let mut ref_tree = PoseidonTree::new(DEFAULT_TREE_DEPTH + 1, ruint::Uint::ZERO);
     let initial_root: U256 = ref_tree.root().into();
 
     let batch_size: usize = 3;
